feat: :sparkles: Implemented shared data services
