+++ src/app/components/sign-in/sign-in.component.ts
@@ -1,10 +1,5 @@
-import { Component, inject } from '@angular/core';
-import {
-  Auth,
-  createUserWithEmailAndPassword,
-  signInWithEmailAndPassword,
-  User,
-} from '@angular/fire/auth';
+import { Component } from '@angular/core';
+import { FirebaseAuthService } from 'src/app/shared/auth/firebase-auth.service';
 
 @Component({
   selector: 'app-sign-in',
@@ -12,19 +7,5 @@ import {
   styleUrls: ['./sign-in.component.scss'],
 })
 export class SignInComponent {
-  auth: Auth = inject(Auth);
-
-  constructor() {}
-
-  signIn(auth: Auth, email: string, password: string) {
-    signInWithEmailAndPassword(auth, email, password)
-      .then((userCredential) => {
-        // Signed in
-        const user = userCredential.user;
-      })
-      .catch((error) => {
-        const errorCode = error.code;
-        const errorMessage = error.message;
-      });
-  }
+  constructor(protected fireAuthService: FirebaseAuthService) {}
 }
