+++ src/app/components/sign-out/sign-out.component.html
@@ -1,2 +1,2 @@
 <p>sign-out works!</p>
-<button (click)="signOut()">Sign Out</button>
+<button (click)="fireAuthService.signOut()">Sign Out</button>
