+++ src/app/components/sign-up/sign-up.component.html
@@ -3,7 +3,16 @@
   <form>
     <input type="email" placeholder="Email" #email required />
     <input type="password" placeholder="Password" #password required />
-    <button type="button" (click)="signUp(auth, email.value, password.value)">
+    <button
+      type="button"
+      (click)="
+        fireAuthService.signUp(
+          fireAuthService.auth,
+          email.value,
+          password.value
+        )
+      "
+    >
       Sign Up
     </button>
   </form>
