+++ src/app/components/user-check/user-check.component.ts
@@ -1,26 +1,15 @@
-import { Component, inject } from '@angular/core';
-import { Auth, User, user } from '@angular/fire/auth';
-import { Subscription } from 'rxjs';
+import { Component, OnDestroy } from '@angular/core';
+import { FirebaseAuthService } from 'src/app/shared/auth/firebase-auth.service';
 
 @Component({
   selector: 'app-user-check',
   templateUrl: './user-check.component.html',
   styleUrls: ['./user-check.component.scss'],
 })
-export class UserCheckComponent {
-  auth: Auth = inject(Auth);
-  user$ = user(this.auth);
-  userSubscription: Subscription;
-  user: User | undefined;
-
-  constructor() {
-    this.userSubscription = this.user$.subscribe((aUser: User | null) => {
-      //handle user state changes here. Note, that user will be null if there is no currently logged in user.
-    });
-  }
+export class UserCheckComponent implements OnDestroy {
+  constructor(protected fireAuthService: FirebaseAuthService) {}
 
   ngOnDestroy() {
-    // when manually subscribing to an observable remember to unsubscribe in ngOnDestroy
-    this.userSubscription.unsubscribe();
+    this.fireAuthService.userSubscription.unsubscribe();
   }
 }
