+++ src/app/models/itinerary-item.ts
@@ -1,0 +1,10 @@
+export interface ItineraryItem {
+  name: string;
+  tag: string;
+  startDate: string;
+  endDate: string;
+  cost: string;
+  startLocation?: string;
+  endLocation?: string;
+  notes?: string;
+}
