+++ src/app/shared/auth/firebase-auth.service.ts
@@ -1,0 +1,53 @@
+import { Injectable, inject } from '@angular/core';
+import {
+  Auth,
+  createUserWithEmailAndPassword,
+  signInWithEmailAndPassword,
+  User,
+  user,
+} from '@angular/fire/auth';
+import { Subscription } from 'rxjs';
+
+@Injectable({
+  providedIn: 'root',
+})
+export class FirebaseAuthService {
+  auth: Auth = inject(Auth);
+  user$ = user(this.auth);
+  userSubscription: Subscription;
+  user: User | undefined;
+
+  constructor() {
+    this.userSubscription = this.user$.subscribe((aUser: User | null) => {
+      //handle user state changes here. Note, that user will be null if there is no currently logged in user.
+    });
+  }
+
+  signIn(auth: Auth, email: string, password: string) {
+    signInWithEmailAndPassword(auth, email, password)
+      .then((userCredential) => {
+        // Signed in
+        const user = userCredential.user;
+      })
+      .catch((error) => {
+        const errorCode = error.code;
+        console.log(error.message);
+      });
+  }
+
+  signOut() {
+    this.auth.signOut();
+  }
+
+  signUp(auth: Auth, email: string, password: string) {
+    createUserWithEmailAndPassword(auth, email, password)
+      .then((userCredential) => {
+        // Signed in
+        const user = userCredential.user;
+      })
+      .catch((error) => {
+        const errorCode = error.code;
+        console.log(error.message);
+      });
+  }
+}
