+++ src/app/shared/store/firebase-store.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { FirebaseStoreService } from './firebase-store.service';
+
+describe('FirebaseStoreService', () => {
+  let service: FirebaseStoreService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(FirebaseStoreService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
